Improve `attribution init` and project namespace guessing
This improves the project namespace guessing to use a canonicalized
form of the package name, with dashes replaced with underscores. It also
greatly improves the tomlkit usage to prevent unnecessary/empty `[tool]`
tables being added to `pyproject.toml`, and makes the final whitespace
more closely match what most people would actually want/expect. Lastly,
it adds a new question for signed tags and includes that option in the
final `pyproject.toml` as well.