Fix #30: use full shortlog for first tag in tree
If there is no preceding tag, the describe command above will just return
the rev hash of the commit preceding the tag, resulting in a bad shortlog.
We can use `git tag -l <name>` to see if it's a real tag or just a hash.
If it's just a rev, then this is the earliest tag in that tree, and we
should just generate a shortlog without a starting ref.
To save shell commands, we assume that any Version-like name is a tag.